pyproject.toml
[project]
name = "streaming-flow-policy"
version = "0.1.0"
description = ""
authors = [
    {name = "Siddharth Ancha", email = "sancha@mit.edu"},
    {name = "Sunshine Jiang", email = "sunsh16e@mit.edu "},
]
dependencies = [
    # =========================================================================
    # Debugging, interactive Python and Jupyter notebooks
    # =========================================================================
    "ipdb",
    "ipykernel",
    "ipywidgets",
    "pyreadline3",
    "tqdm",
    "jupyterlab",
    "jupyter-book",

    # =========================================================================
    # Main dependencies
    # =========================================================================
    "mediapy",
    "opencv-python",
    "drake",
    "matplotlib",
    "termcolor",
    "scipy",
    "scikit-image",
    "scikit-video",
    "torch",
    "torchvision",
    "torchdyn",
    "diffusers",
    "zarr",
    "numcodecs",
    "pygame",
    "pymunk",
    "gdown",
    "gym",
    "shapely",
]
requires-python = "<3.12,>=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
